pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SMSROOMBackend"
version = "0.1.0"
description = "SMS Room REST API"
authors = [
    {name = "karrad/karrad1201", email = "karad120109@gmail.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "pydantic>=2.0.0",
    "pydantic[email]",
    "aiomysql",
    "sqlalchemy>=2.0.0",
    "python-dotenv",
    "bcrypt",
    "pyjwt",
    "alembic",
    "asyncpg",
    "httpx",
    "python-multipart"
]


[project.optional-dependencies]
dev = [
    "pytest[asyncio]>=6.0",
    "aiosqlite",
    "httpx"
]
